feat: add color option for small widget tweak: emporarily disable widget customization entry point feat: add clearer text for average upload time
