+++ chafenqi/View/v2/Settings/Settings.swift
@@ -72,15 +72,15 @@ struct Settings: View {
                 } label: {
                     Text("排序")
                 }
-                NavigationLink {
-                    if user.isPremium {
-                        SettingsWidgetConfig(user: user)
-                    } else {
-                        NotPremiumView()
-                    }
-                } label: {
-                    Text("小组件")
-                }
+//                NavigationLink {
+//                    if user.isPremium {
+//                        SettingsWidgetConfig(user: user)
+//                    } else {
+//                        NotPremiumView()
+//                    }
+//                } label: {
+//                    Text("小组件")
+//                }
             } header: {
                 Text("主页")
             }
